+++ go.mod
@@ -4,11 +4,11 @@ go 1.17
 
 require (
 	github.com/Clinet/discordgo-embed v0.0.0-20220113222025-bafe0c917646
-	github.com/bwmarrin/discordgo v0.23.2
+	github.com/bwmarrin/discordgo v0.24.0
 )
 
 require (
-	github.com/clinet/discordgo-embed v0.0.0-20220113222025-bafe0c917646 // indirect
-	github.com/gorilla/websocket v1.4.0 // indirect
-	golang.org/x/crypto v0.0.0-20181030102418-4d3f4d9ffa16 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	golang.org/x/crypto v0.0.0-20210421170649-83a5a9bb288b // indirect
+	golang.org/x/sys v0.0.0-20201119102817-f84b799fce68 // indirect
 )
